diffreg-sbir/requirements.txt
torch>=1.13
torchvision>=0.14
numpy
pyyaml
tqdm
scikit-learn
pillow
matplotlib
faiss-cpu